update the code and fixed all errors

+++ src/app/modules/users/users.service.ts
@@ -3,18 +3,11 @@ import { User_Type, userOrderType } from './users.interface';
 import { User_Model } from './users.model';
 
 const createUserintoDB = async (user: User_Type) => {
-
-  //================== instrance =================
-  const userInstrance = new User_Model(user);
-
-  if (await userInstrance.isUserExists(user.userId)) {
-    throw new Error(
-      'The user already exists!!!, Please check your userId. It needs to be unique...',
-    );
+  if (await User_Model.isUserExists(user.userId)) {
+    throw new Error('The user already exists!, Duplicate userId');
   }
 
-
-  const result = await userInstrance.save(); //built in instrance method
+  const result = await User_Model.create(user);
 
   return result;
 };
@@ -25,123 +18,83 @@ const getAllUserintoDB = async () => {
 };
 
 const getSingleUserintoDB = async (userId: string) => {
-  try {
-    const id = Number(userId);
+  const id = Number(userId);
+
+  if (await User_Model.isUserExists(id)) {
     const result = await User_Model.findOne({ userId: id });
-
-    
-
-    if (!result) {
-      throw new Error("User not Found!");
-    }
     return result;
-
+  } else {
+    throw new Error('The user not fiund!');
   }
-  catch (error: any) {
-    console.log(error)
-    throw new Error(error)
-  }
-
 };
 
 const deleteSingleUserintoDB = async (userId: string) => {
-  try {
-    const id = Number(userId);
+  const id = Number(userId);
+
+  if (await User_Model.isUserExists(id)) {
     const result = await User_Model.deleteOne({ userId: id });
-    if (result.deletedCount === 0) {
-      throw Error("User not Found!!!");
-    }
     return result;
-  } catch (error) {
-    console.log(error)
-    throw Error("User not found!!")
+  } else {
+    throw new Error('User not found!');
   }
 };
 
 const updateUserIntoDB = async (userId: string, payload: User_Type) => {
-  try {
-    const id = Number(userId);
+  const id = Number(userId);
+
+  if (await User_Model.isUserExists(id)) {
     const result = await User_Model.findOneAndUpdate({ userId: id }, payload, {
       new: true,
     });
-    if (!result) {
-      throw new Error("User not Found!");
-    }
     return result;
-  } catch (error: any) {
-    console.log(error)
-    throw new Error(error)
+  } else {
+    throw new Error('User not Found!');
   }
-
 };
 
 const createOrderIntoDB = async (userId: string, payload: userOrderType) => {
-  try {
+  const id = Number(userId);
 
-    const id = Number(userId);
+  if (await User_Model.isUserExists(id)) {
     const result = await User_Model.findOneAndUpdate(
       { userId: id },
       { $push: { orders: payload } },
-      { new: true });
-    if (!result) {
-      throw new Error("User not Found!!!");
-    }
+      { new: true },
+    );
     return result;
-  } catch (error: any) {
-    console.log(error)
-    throw new Error(error)
+  } else {
+    throw new Error('User not Found!');
   }
-
 };
 
 const getAllOrderIntoDB = async (userId: string) => {
-  try {
+  const id = Number(userId);
 
-    const id = Number(userId);
+  if (await User_Model.isUserExists(id)) {
     const result = await User_Model.findOne({ userId: id });
-
-    if (!result) {
-      throw new Error("User not Found!!!");
-    }
     return result?.orders;
-
+  } else {
+    throw new Error('User not Found!');
   }
-  catch (error: any) {
-    console.log(error)
-    throw new Error(error)
-  }
-
 };
 
-
 const getTotalPriceOfOrderIntoDB = async (userId: string) => {
-  try {
+  const id = Number(userId);
 
-    const id = Number(userId);
+  if (await User_Model.isUserExists(id)) {
     const result = await User_Model.findOne({ userId: id });
-
-    if (!result) {
-      throw new Error("User not Found!!!");
-    }
 
     let totalPrice: number = 0;
 
-    result.orders.forEach((element) => {
+    result?.orders.forEach((element) => {
       totalPrice += element.price * element.quantity;
     });
 
-
     return parseFloat(totalPrice.toFixed(2));
-
+  } else {
+    throw new Error('User not Found!');
   }
-  catch (error: any) {
-    console.log(error)
-    throw new Error(error)
-  }
-
 };
-
-
 
 export const User_Services = {
   createUserintoDB,
@@ -151,9 +104,8 @@ export const User_Services = {
   updateUserIntoDB,
   createOrderIntoDB,
   getAllOrderIntoDB,
-  getTotalPriceOfOrderIntoDB
+  getTotalPriceOfOrderIntoDB,
 };
-
 
 // const result = await User.findByIdAndUpdate({ _id: id }, payload, {
 //    new: true,
